Patch to sample safe primes in release builds (#35)
Sample safe primes when not running tests

This commit breaks the safe prime sampling function into a "test" version which reads primes from a file for speed (and can only be called from tests) and a production version which actually samples the safe primes. Tests were modified to use the test version.