Parameterize all random functions with an RNG (#88)
This PR addresses #29 by modifying existing library functions so that every function which involves sampling values nondeterministically is parameterized with an rng to do this sampling. This also addresses #28 as the library itself should not call any of its own rngs, thus making tests deterministic once an rng is initialized.

More specifically, this PR adds a test-only function called get_test_rng() to utils.rs which samples a random seed, prints the seed to stderr, and returns an rng seeded with this seed. This function is called only once per test and is only called in functions with the #[test] decorator.