Use `PaillierCiphertext` in ZKPs (#104)
Use `PaillierCiphertext` in ZKPs

This commit replaces the use of `BigNumber` with `PaillierCiphertext`
where needed in the ZKPs. It also adds a `to_bytes` method to
`PaillierCiphertext` to avoid needing to explicitly call the
`BigNumber::to_bytes` method. Further hiding of the underlying
`BigNumber` value will be made once Issue #64 is addressed.

Closes #63.