Add `PaillierNonce` wrapper to API (#106)
Add `PaillierNonce` wrapper to API

This PR adds a `PaillierNonce` type which wraps a `BigNumber`, and
utilizes this new type where needed in the various ZKPs. The PR
also removes an unnecessary check that the generated nonce is non-zero
(the check is unnecessary as we already to such a check during nonce generation).
Finally, the PR hides the internal nonce representation behind an `inner` method.