Add nonce masking operation to `PaillierEncryptionKey` (#124)
This MR adds a new `PaillierEncryptionKey::mask` method, which
takes two nonces and produces a masked nonce. It also threads the
`MaskedNonce` type through the ZKPs, although currently we still use
`.0` to extract the underlying `BigNumber`. This will change once we
add some additional functionality for operating on `MaskedNonce`s.