Address issue #40 and add ZKStarN API
Adds a ZKStarBuilder that defines the modulus of the multiplicative group and allows serializing elements of ZKStarN without duplicating the modulus.
Adds ZKStarN which defines elements of the multiplicative group. These elements are correctly members of the group by construction and can only be serialized.
Adds ZKStarNUnverified which defines dserialized values that claim to be members of the multiplicative group mod N and that need to be verified before constructing members of ZKStarN.