consolidate errors (#173)
This PR attempts to consolidate some error types, particularly around proof generation, addressing issue #158 .

For this PR I kept the verify_err! macro in place so that a later PR can switch it to a logging one.

For parts of the code made less descriptive by switching the error type, I left a comment indicating the reason for the failure. Once logging is done, this is the text which should be logged