Add meaningful `Input` type to `PresignParticipant` (#245)
Add meaningful `Input` type to `PresignParticipant`

This commit adds an `Input` type to `PresignParticipant` that is more
meaningful than just using `Storage`. In particular, it contains the
necessary auxinfo and keygen information for `PresignParticipant` to
operate. This `Input` can be constructed in `process_single_message`
once `auxinfo` and `keygen` are completed.