Add `status` method to `ProtocolParticipant` trait (#260)
* Add `is_done` method to `ProtocolParticipant` trait

This allows a caller to tell when a protocol has successfully
completed. In addition, this commit adds a `status` enum to each
`ProtocolParticipant` instantiation to track protocol status.
(Currently just `Initialized` and `TerminatedSuccessfully` is used.)

* Replace `is_done` with `status` in `ProtocolParticipant`

Rather than restricting the visibility of `status` to only if the
protocol is done or not, use a `status` method to return the full
`status`, allowing the caller to decide what to do with it.

* Check protocol status in full protocol test

This requires exposing the status in `Participant`, which we probably
should have anyways.

* Fix broadcast status

The status now includes a vector of participants that have completed a
broadcast protocol. We can then check termination by checking that the
number of participants that have completed a broadcast equals the
total number of participants.