Audit of auxinfo protocol (#298)
Comments/auditing info related to the auxinfo protocol

This commit also fixes a bug in the proof! We were not using the
verifier's setup parameters for the ZKPs. This commit fixes that,
although it does result in slower performance (since we need to
generate a fresh proof for each participant).