288-remove (#409)
Adds a new remove() function to the Storage API, which removes an item from storage and returns it to the call. Replaces a few end-of-protocol calls to retrieve with remove, but only where there is no chance that the item will be needed again. Replacing other instances of retrieve with remove will require more care, especially when considering the possibility of ProtocolErrors

Closes #288